requirements.txt
google-api-python-client
pandas
python-dotenv
pytrends
fastapi
uvicorn
schedule